refactor(frontend): extract useSqlCompletion composable (P2-2 F6)
Move the SQL-editor completion data source (schema/table/column getters,
loaders, remote search, per-tab completion context) out of DataStudioNew.vue
into composables/useSqlCompletion.js. The shared catalog caches
(schemaStore/tableStore/columnStore) and loaders (loadTables/
activateDatasource) stay component-owned and are injected (loaders as lazy
wrappers), to be unified when the catalog tree is extracted (F7).

Verified: mount smoke green, npm run build (refs resolve), npm run lint
(0 errors); orphan check clean. Functions are verbatim-moved.
Refs docs/plans/2026-06-18-frontend-megacomponent-split-plan.md (F6).
